WinGui: - The scan log will no longer be written to disk if the log is over 100MB. An error message will be thrown. - The encode will be terminated if the log file grows beyond 100MB and an error message will be thrown. - Fixed an issue with the exception window not appearing when called from a worker thead. - Put a limit on the number of exception windows that can appear during a session of HandBrake. (set to 30)
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@3554 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ win/C#/HandBrake.ApplicationServices/Services/Scan.cs
@@ -28,7 +28,7 @@ namespace HandBrake.ApplicationServices.Services
         /// <summary>
         /// The Error Service
         /// </summary>
-        private IErrorService errorService;
+        private readonly IErrorService errorService;
 
         /// <summary>
         /// A Lock object
@@ -138,7 +138,7 @@ namespace HandBrake.ApplicationServices.Services
         {
             try
             {
-                if (hbProc != null)
+                if (hbProc != null && !hbProc.HasExited)
                     hbProc.Kill();
             }
             catch (Exception ex)
@@ -205,12 +205,21 @@ namespace HandBrake.ApplicationServices.Services
                 this.SouceData = DVD.Parse(this.readData);
 
                 // Write the Buffer out to file.
-                StreamWriter scanLog = new StreamWriter(dvdInfoPath);
-                scanLog.WriteLine(Logging.CreateCliLogHeader(null));
-                scanLog.Write(this.readData.Buffer);
-                scanLog.Flush();
-                scanLog.Close();
-                logBuffer.AppendLine(this.readData.Buffer.ToString());
+                using (StreamWriter scanLog = new StreamWriter(dvdInfoPath))
+                {
+                    // Only write the log file to disk if it's less than 100MB.
+                    if (this.readData.Buffer.Length < 100000000)
+                    {
+                        scanLog.WriteLine(Logging.CreateCliLogHeader(null));
+                        scanLog.Write(this.readData.Buffer);
+                        logBuffer.AppendLine(this.readData.Buffer.ToString());
+                    }
+                    else
+                    {
+                        throw new Exception(
+                            "The Log file has not been written to disk as it has grown above the 100MB limit. This indicates there was a problem during the scan process.");
+                    }
+                }
 
                 IsScanning = false;
 
@@ -219,7 +228,12 @@ namespace HandBrake.ApplicationServices.Services
             }
             catch (Exception exc)
             {
-                errorService.ShowError("frmMain.cs - scanProcess() Error", exc.ToString());
+                this.Stop();
+
+                errorService.ShowError("An error has occured during the scan process.", exc.ToString());
+
+                if (this.ScanCompleted != null)
+                    this.ScanCompleted(this, new EventArgs());   
             }
         }
